move set_mandatory_receiver_affirmation tests to settlement_test.rs
