+++ pallets/runtime/tests/src/asset_test.rs
@@ -39,15 +39,15 @@ use polymesh_primitives::calendar::{CalendarPeriod, CalendarUnit, FixedOrVariabl
 use polymesh_primitives::checkpoint::{NextCheckpoints, ScheduleCheckpoints, ScheduleId};
 use polymesh_primitives::constants::currency::ONE_UNIT;
 use polymesh_primitives::settlement::{
-    AffirmationRequirement, InstructionId, Leg, SettlementType, VenueDetails, VenueId, VenueType,
+    InstructionId, Leg, SettlementType, VenueDetails, VenueId, VenueType,
 };
 use polymesh_primitives::statistics::StatType;
 use polymesh_primitives::statistics::{Stat1stKey, Stat2ndKey};
-use polymesh_primitives::traits::{AffirmationFnTrait, AssetFnTrait};
+use polymesh_primitives::traits::AssetFnTrait;
 use polymesh_primitives::{
-    AssetHolder, AssetHolderKind, AssetIdentifier, AssetPermissions, AuthorizationData, Document,
-    DocumentId, Fund, FundDescription, IdentityId, Memo, Moment, NFTCollectionKeys, Permissions,
-    PortfolioId, PortfolioKind, PortfolioName, PortfolioNumber, Signatory, Ticker, WeightMeter,
+    AssetHolderKind, AssetIdentifier, AssetPermissions, AuthorizationData, Document, DocumentId,
+    Fund, FundDescription, IdentityId, Memo, Moment, NFTCollectionKeys, Permissions, PortfolioId,
+    PortfolioKind, PortfolioName, PortfolioNumber, Signatory, Ticker, WeightMeter,
 };
 use sp_keyring::Sr25519Keyring;
 
@@ -1800,51 +1800,6 @@ fn remove_asset_pre_approval() {
         assert!(!PreApprovedAsset::<TestStorage>::get(alice.did, asset_id));
         assert!(!AssetsExemptFromAffirmation::<TestStorage>::get(asset_id));
         assert!(!Asset::skip_asset_affirmation(&alice.did, &asset_id));
-    });
-}
-
-#[test]
-fn set_mandatory_receiver_affirmation() {
-    ExtBuilder::default().build().execute_with(|| {
-        let alice = User::new(Sr25519Keyring::Alice);
-        let asset_id = AssetId::new([0; 16]);
-        let alice_holder: AssetHolder = PortfolioId::default_portfolio(alice.did).into();
-
-        // Default: no mandatory receiver affirmation.
-        assert!(!Settlement::identity_requires_affirmation(&alice.did));
-        // Receiver affirmation is skipped by default.
-        assert!(Asset::skip_asset_holder_affirmation(&alice_holder, &asset_id).unwrap());
-
-        // Opt-in to mandatory receiver affirmation.
-        assert_ok!(Settlement::set_mandatory_receiver_affirmation(
-            alice.origin(),
-            AffirmationRequirement::Required
-        ));
-        assert!(Settlement::identity_requires_affirmation(&alice.did));
-        // Now receiver affirmation is required.
-        assert!(!Asset::skip_asset_holder_affirmation(&alice_holder, &asset_id).unwrap());
-
-        // Pre-approve a specific asset — overrides mandatory receiver affirmation.
-        assert_ok!(Asset::pre_approve_asset(alice.origin(), asset_id));
-        assert!(Asset::skip_asset_holder_affirmation(&alice_holder, &asset_id).unwrap());
-
-        // Remove pre-approval — mandatory receiver affirmation works.
-        assert_ok!(Asset::remove_asset_pre_approval(alice.origin(), asset_id));
-        assert!(!Asset::skip_asset_holder_affirmation(&alice_holder, &asset_id).unwrap());
-
-        // Global asset exemption overrides mandatory receiver affirmation.
-        assert_ok!(Asset::exempt_asset_affirmation(root(), asset_id));
-        assert!(Asset::skip_asset_holder_affirmation(&alice_holder, &asset_id).unwrap());
-        assert_ok!(Asset::remove_asset_affirmation_exemption(root(), asset_id));
-
-        // Opt out of mandatory receiver affirmation.
-        assert_ok!(Settlement::set_mandatory_receiver_affirmation(
-            alice.origin(),
-            AffirmationRequirement::Automatic
-        ));
-        assert!(!Settlement::identity_requires_affirmation(&alice.did));
-        // Affirmation is skipped again.
-        assert!(Asset::skip_asset_holder_affirmation(&alice_holder, &asset_id).unwrap());
     });
 }
